+++ test/unit/Intelligence/DetectorRegistryTest.php
@@ -79,6 +79,7 @@ class DetectorRegistryTest extends TestCase {
 				'vulnerability-probes',
 				'html-injection',
 				'php-probes',
+				'legacy-endpoints',
 			),
 			Detector_Registry::keys()
 		);
@@ -88,7 +89,7 @@ class DetectorRegistryTest extends TestCase {
 		Detector_Registry::register_defaults();
 		Detector_Registry::register_defaults();
 
-		$this->assertCount( 12, Detector_Registry::keys() );
+		$this->assertCount( 13, Detector_Registry::keys() );
 	}
 
 	public function test_reset_allows_register_defaults_to_repopulate(): void {
@@ -99,7 +100,7 @@ class DetectorRegistryTest extends TestCase {
 
 		Detector_Registry::register_defaults();
 
-		$this->assertCount( 12, Detector_Registry::keys() );
+		$this->assertCount( 13, Detector_Registry::keys() );
 	}
 }